// types/carousel.ts
export interface TextAnalysisData {
    fullOutput(fullOutput: any, arg1: null, arg2: number): import("react").ReactNode
    mainTitle: string
    author: string
    importantFacts: string[]
    opposingForces: string[]
    notableQuotes: string[]
  }
  
  export interface CarouselData {
    Title: string
    Subtitle: string
    Text01: string
    Text02: string
    Text03: string
    Hashtag: string
  }
  
  